Bitstring fix (Addresses CQCL/tket#1102)
The conditional value of the control state falls through the OR
condition when the control state is defined but 0. This commit addresses
this by replacing the evaluation of the controlState with a nullish
coalescing operator.

An example is demonstrated in the storybook by adding two clones of the
leftmost control gate (which itself has no control_state set), one
having control_state=0 and the other having control_state=1.